Consolidate autojoin/summon code
Autojoining and autosummoning is now all in one function.  The bot
should now autosummon if it finds the owner in a voice channel and
there's not an autojoin channel set for that server.

Additionally, the bot will now resume voice connections if possible.
This means if it sees itself in a voice channel it'll join that voice
channel again so it won't appear in it without being in it.