Fix bandcamp album issue (#1835)
After creating your pull request, tick these boxes if they are applicable to you.

- [x] I have tested my changes against the `review` branch (the latest developmental version), and this pull request is targeting that branch as a base
- [x] I have tested my changes on Python 3.5/3.6

----

### Description
Fix #1827

This is the nicer version of the fix I proposed in the server. Wrapping the lock with try block and catch the exception is not really encourage readability. Here we try to get the actual URL first and then go down the line which is more intuitive.


### Related issues (if applicable)